requirements.txt
streamlit>=1.28.0
dspy-ai>=2.4.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
firecrawl-py>=1.0.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
anthropic>=0.25.0
tiktoken>=0.5.0
pydantic>=2.0.0
typing-extensions>=4.8.0